refactor: Update NiveauCompetence model with fillable property
The NiveauCompetence model has been updated to include the `fillable` property. This property specifies the fields that are allowed to be mass assigned when creating or updating a NiveauCompetence instance. The `nom` and `description` fields have been added to the `fillable` array.

Refactor the NiveauCompetence model to include the `fillable` property and add the `nom` and `description` fields to the `fillable` array.